final cleanup

+++ mflowgen/common/rtl/gen_rtl.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-START_DIR=$(pwd)
 # Hierarchical flows can accept RTL as an input from parent graph
 if [ -f ../inputs/design.v ]; then
   echo "Using RTL from parent graph"
@@ -27,7 +26,6 @@ else
 
     # Use aha docker container for all dependencies
     if [ $use_container == True ]; then
-        set -x
       # Clone AHA repo
       git clone https://github.com/StanfordAHA/aha.git
       cd aha
@@ -37,57 +35,28 @@ else
       # Prune docker images...
       yes | docker image prune -a --filter "until=6h" --filter=label='description=garnet' || true
 
-##############################################################################
-which_container=cst
-#which_container=latest
+      ##############################################################################
+      # steveri 02/2021 - Original code only supported container "latest";
+      # new code (below) allows use of any container
 
-# if [ "$which_container" == "latest" ]; then
-# 
-#    # ORIGINAL CODE
-#       # pull docker image from docker hub
-#       docker pull stanfordaha/garnet:latest
-# 
-#       # run the container in the background and delete it when it exits
-#       # (this will print out the name of the container to attach to)
-#       container_name=$(aha docker)
-#       echo "container-name: $container_name"
-# else
-# 
-#    # TEMPORARY CST-CHECK HACK
-#       # pull docker image from docker hub
-#       docker pull stanfordaha/garnet:cst
-# 
-#       # run the container in the background and delete it when it exits
-#       # (this will print out the name of the container to attach to)
-#       container_name=cst
-#       echo "container-name: $container_name"
-#       # mount the /cad and name it, also run it as a daemon in background
-#       docker run -id --name ${container_name} --rm -v /cad:/cad stanfordaha/garnet:cst bash
-# fi
-##############################################################################
+      which_container=cst
+      #which_container=latest
 
-
-
-
-# So...does this even work? No it does not!
-
-   # ORIGINAL CODE
       # pull docker image from docker hub
-      docker pull stanfordaha/garnet:cst
+      docker pull stanfordaha/garnet:${which_container}
 
-#       # run the container in the background and delete it when it exits
-#       # (this will print out the name of the container to attach to)
-#       container_name=$(aha docker)
-#       echo "container-name: $container_name"
-
-
-      container_name=cst
+      if [ "$which_container" == "latest" ]; then
+          # run the container in the background and delete it when it exits
+          # ("aha docker" will print out the name of the container to attach to)
+          container_name=$(aha docker)
+      else
+          # run the container in the background and delete it when it exits (--rm)
+          # mount /cad and name it, and run container as a daemon in background
+          container_name=${which_container}
+          docker run -id --name ${container_name} --rm -v /cad:/cad stanfordaha/garnet:cst bash
+      fi
       echo "container-name: $container_name"
-      # mount the /cad and name it, also run it as a daemon in background
-      docker run -id --name ${container_name} --rm -v /cad:/cad stanfordaha/garnet:cst bash
-
-
-
+      ##############################################################################
 
       if [ $use_local_garnet == True ]; then
         docker exec $container_name /bin/bash -c "rm -rf /aha/garnet"
@@ -100,9 +69,10 @@ which_container=cst
       # run garnet.py in container and concat all verilog outputs
       docker exec $container_name /bin/bash -c \
         '# Single-quote regime
-         echo sq
-         # Func to check python package creds
-         set -x
+
+         ###########################################################################
+         # Func to check python package creds (Added 02/2021 as part of cst vetting)
+
          function checkpip {
              # Example: checkpip ast.t "peak "
              #   ast-tools              0.0.18    /usr/local/venv_garnet/src/ast-tools
@@ -118,19 +88,15 @@ which_container=cst
              done
          }'"
          # Double-quote regime
-         echo dq
          source /aha/bin/activate
 
-         # Build garnet verilog; check to see that the right packages get used
+         # Build garnet verilog; check and double-check cst packages
          echo 'PIPCHECK1-BEFORE'; checkpip ast.t magma 'peak '
-         exit
-
-
          aha garnet $flags;
          echo 'PIPCHECK2-AFTER';  checkpip ast.t magma 'peak '
-         exit
+
          cd garnet
-         if [ -d \"genesis_verif\" ]; then
+         if [ -d 'genesis_verif' ]; then
            cp garnet.v genesis_verif/garnet.v
            cat genesis_verif/* >> design.v
          else
@@ -184,16 +150,3 @@ which_container=cst
     cd $current_dir
   fi
 fi
-
-echo ENDGAME
-set -x
-pwd
-current_dir=$(pwd)
-cd $START_DIR
-  # cp outputs/design.v /tmp/design.v.$$
-  # cp mflowgen-run.log /tmp/log.$$
-  pwd; ls -l outputs/design.v mflowgen-run.log
-  cp outputs/design.v /tmp/design.v.${which_container}.deleteme$$
-  cp mflowgen-run.log /tmp/cstlog.${which_container}.deleteme$$
-cd $current_dir
-set +x
